backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
twilio==8.5.0
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.0
anthropic==0.25.0
firebase-admin==6.5.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
